Move lgssm model code to lgssm/models.py
